Move the load-document workflow out of parse/ into its own directory
parse/ mixed the load-document Workflow (orchestrator + steps) with
unrelated, reusable parsing helpers (insert-and-fasten, parse-configuration,
parse-vendors, build-checks) that aren't workflow-specific — insert-and-fasten
is called directly from routes/insertables.ts, for instance. This made it
unclear where a future workflow would live, and made the load-document /
load-document-steps split feel like an arbitrary same-folder naming
convention rather than a real boundary.

load-document.ts and load-document-steps.ts now live in their own
src/backend/load-document-workflow/ directory (steps.ts, dropping the
redundant folder-scoped prefix), sitting as a sibling to routes/,
onshape-api/, and parse/. A future workflow (e.g. thumbnail loading) follows
the same <name>-workflow/ shape. parse/ now holds only the reusable parsing
helpers it was always mostly made of.
